UCSD MAE 150 - CAD Analysis for MATLAB
The following sets of scripts are developed in MATLAB based on the UCSD
MAE150 class curriculum. The exercises were designed to allow a
comprehensive understanding of the built-in functions in MATLAB
scripting, as well as application of basic physics principles.
Follow the decriptions on the written homework .docx file to get a more
descriptive idea of the purposes of the scripts.